+++ src/main/java/frc/robot/Main.java
@@ -19,7 +19,7 @@ public final class Main {
      * <p>If you change your main robot class, change the parameter type.
      */
     public static void main(String... args) {
-        RobotBase.startRobot(Robot::new);
+        RobotBase.startRobot(RobotManager::new);
     }
 
 }
