+++ src/main/java/frc/robot/Robot.java
@@ -5,68 +5,50 @@
 package frc.robot;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.CommandScheduler;
-import frc.robot.constants.RobotConstants;
+import edu.wpi.first.wpilibj2.command.InstantCommand;
+import frc.robot.poseestimation.poseestimator.PoseEstimator;
+import frc.robot.subsystems.swerve.Swerve;
+import frc.robot.subsystems.swerve.SwerveConstants;
+import frc.robot.subsystems.swerve.SwerveState;
+import frc.utils.DriverStationUtils;
 import frc.utils.pathplannerutils.PathPlannerUtils;
-import frc.robot.simulation.SimulationManager;
-import frc.utils.battery.BatteryUtils;
-import frc.utils.ctre.BusStatus;
-import frc.utils.cycletime.CycleTimeUtils;
-import frc.utils.logger.LoggerFactory;
-import org.littletonrobotics.junction.LoggedRobot;
 
 /**
- * The VM is configured to automatically run this class, and to call the functions corresponding to
- * each mode, as described in the TimedRobot documentation. If you change the name of this class or
- * the package after creating this project, you must also update the build.gradle file in the
- * project.
+ * This class is where the bulk of the robot should be declared. Since Command-based is a
+ * "declarative" paradigm, very little robot logic should actually be handled in the {@link RobotManager}
+ * periodic methods (other than the scheduler calls). Instead, the structure of the robot (including
+ * subsystems, commands, and trigger mappings) should be declared here.
  */
-public class Robot extends LoggedRobot {
+public class Robot {
 
-    private Command autonomousCommand;
+    public static final Swerve SWERVE = new Swerve(); // Must be before POSE_ESTIMATOR
 
-    private RobotContainer robotContainer;
+    public static final PoseEstimator POSE_ESTIMATOR = new PoseEstimator();
 
-    @Override
-    public void robotInit() {
-        if (RobotConstants.ROBOT_TYPE.isReplay()) {
-            setUseTiming(false); // run as fast as possible
-        }
-        LoggerFactory.initializeLogger();
-        BatteryUtils.scheduleLimiter(); // Using RobotConstants.BATTERY_LIMITER_ENABLE, disable with it!
-        PathPlannerUtils.startPathPlanner();
-
-        robotContainer = new RobotContainer();
+    public Robot() {
+        buildPathPlannerForAuto();
+        configureBindings();
     }
 
-    @Override
-    public void autonomousInit() {
-        autonomousCommand = robotContainer.getAutonomousCommand();
-
-        if (autonomousCommand != null) {
-            autonomousCommand.schedule();
-        }
+    private void buildPathPlannerForAuto() {
+        // Register commands...
+        PathPlannerUtils.configurePathPlanner(
+                Robot.POSE_ESTIMATOR::getCurrentPose,
+                Robot.POSE_ESTIMATOR::resetPose,
+                Robot.SWERVE::getSelfRelativeVelocity,
+                (speeds) -> Robot.SWERVE.driveByState(speeds, SwerveState.DEFAULT_PATH_PLANNER),
+                SwerveConstants.HOLONOMIC_PATH_FOLLOWER_CONFIG,
+                DriverStationUtils::isRedAlliance,
+                Robot.SWERVE
+        );
     }
 
-    @Override
-    public void teleopInit() {
-        if (autonomousCommand != null) {
-            autonomousCommand.cancel();
-        }
+    private void configureBindings() {
+        JoysticksBindings.configureBindings();
     }
 
-    @Override
-    public void robotPeriodic() {
-        CycleTimeUtils.updateCycleTime(); // Better to be first
-        CommandScheduler.getInstance().run();
-        RobotContainer.POSE_ESTIMATOR.periodic();
-        BusStatus.logChainsStatuses();
-        BatteryUtils.logStatus();
-    }
-
-    @Override
-    public void simulationPeriodic() {
-        SimulationManager.updateRegisteredSimulations();
+    public Command getAutonomousCommand() {
+        return new InstantCommand();
     }
 
 }
