--- src/main/java/frc/robot/RobotContainer.java
@@ -1,76 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot;
-
-import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.InstantCommand;
-import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
-import edu.wpi.first.wpilibj2.command.button.Trigger;
-import frc.robot.poseestimation.poseestimator.PoseEstimator;
-import frc.robot.subsystems.swerve.Swerve;
-import frc.robot.subsystems.swerve.SwerveConstants;
-import frc.robot.subsystems.swerve.SwerveState;
-import frc.utils.DriverStationUtils;
-import frc.utils.pathplannerutils.PathPlannerUtils;
-
-/**
- * This class is where the bulk of the robot should be declared. Since Command-based is a
- * "declarative" paradigm, very little robot logic should actually be handled in the {@link Robot}
- * periodic methods (other than the scheduler calls). Instead, the structure of the robot (including
- * subsystems, commands, and trigger mappings) should be declared here.
- */
-public class RobotContainer {
-
-    public static final Swerve SWERVE = new Swerve(); // Must be before POSE_ESTIMATOR
-
-    public static final PoseEstimator POSE_ESTIMATOR = new PoseEstimator();
-
-    /**
-     * The container for the robot. Contains subsystems, OI devices, and commands.
-     */
-    public RobotContainer() {
-        buildPathPlannerForAuto();
-        configureBindings();
-    }
-
-    /**
-     * Use this method to define your trigger->command mappings. Triggers can be created via the
-     * {@link Trigger#Trigger(java.util.function.BooleanSupplier)} constructor with an arbitrary
-     * predicate, or via the named factories in {@link
-     * edu.wpi.first.wpilibj2.command.button.CommandGenericHID}'s subclasses for {@link
-     * CommandXboxController Xbox}/{@link edu.wpi.first.wpilibj2.command.button.CommandPS4Controller
-     * PS4} controllers or {@link edu.wpi.first.wpilibj2.command.button.CommandJoystick Flight
-     * joysticks}.
-     */
-    private void configureBindings() {
-        JoysticksBindings.configureBindings();
-    }
-
-    /**
-     * Use this to pass the autonomous command to the main {@link Robot} class.
-     *
-     * @return the command to run in autonomous
-     */
-    public Command getAutonomousCommand() {
-        return new InstantCommand();
-    }
-
-
-    private void buildPathPlannerForAuto() {
-        // Register commands...
-        PathPlannerUtils.configurePathPlanner(
-                RobotContainer.POSE_ESTIMATOR::getCurrentPose,
-                RobotContainer.POSE_ESTIMATOR::resetPose,
-                RobotContainer.SWERVE::getSelfRelativeVelocity,
-                (speeds) -> RobotContainer.SWERVE.driveByState(speeds, SwerveState.DEFAULT_PATH_PLANNER),
-                SwerveConstants.HOLONOMIC_PATH_FOLLOWER_CONFIG,
-                DriverStationUtils::isRedAlliance,
-                RobotContainer.SWERVE
-        );
-    }
-
-
-
-}
