+++ src/main/java/frc/robot/subsystems/swerve/gyro/gyrointerface/SwerveGyroFactory.java
@@ -1,9 +1,9 @@
 package frc.robot.subsystems.swerve.gyro.gyrointerface;
 
 import frc.robot.constants.RobotConstants;
+import frc.robot.subsystems.swerve.gyro.pigeon2.Pigeon2Gyro;
 import frc.robot.subsystems.swerve.gyro.replay.ReplaySwerveGyro;
 import frc.robot.subsystems.swerve.gyro.simulation.SimulationSwerveGyro;
-import frc.robot.subsystems.swerve.gyro.pigeon2.Pigeon2Gyro;
 
 public class SwerveGyroFactory {
