+++ src/main/java/frc/robot/subsystems/swerve/modules/moduleinterface/ModuleFactory.java
@@ -2,8 +2,8 @@ package frc.robot.subsystems.swerve.modules.moduleinterface;
 
 import frc.robot.constants.RobotConstants;
 import frc.robot.subsystems.swerve.modules.ModuleUtils;
-import frc.robot.subsystems.swerve.modules.replay.ReplayModule;
 import frc.robot.subsystems.swerve.modules.mk4i.MK4IModule;
+import frc.robot.subsystems.swerve.modules.replay.ReplayModule;
 import frc.robot.subsystems.swerve.modules.simulation.SimulationModule;
 
 public class ModuleFactory {
